Add Klarna request to retrieve widget

+++ src/Api/Checkout/ShipmentLine.php
@@ -61,7 +61,7 @@ class ShipmentLine extends AbstractLineItem
         $this->name = $name;
         $this->quantity = $shipment->getShippingUnitCount();
         $this->quantityUnit = 'pcs';
-        $this->unitPrice = $shippingCharge/$shipment->getShippingUnitCount();
+        $this->unitPrice = (int) ($shippingCharge / $shipment->getShippingUnitCount());
         $this->taxRate = $shippingTaxRate;
         $this->totalAmount = $shippingCharge;
         $this->totalDiscountAmount = 0;
